+++ system/src/Grav/Framework/Flex/Pages/FlexPageObject.php
@@ -193,13 +193,14 @@ class FlexPageObject extends FlexObject implements PageInterface, MediaManipulat
 
         return $key;
     }
+
     /**
      * {@inheritdoc}
      * @see FlexObjectInterface::getCacheKey()
      */
     public function getCacheKey(): string
     {
-        return $this->hasKey() ? $this->getTypePrefix() . $this->getFlexType() . '.' . $this->getKey() . $this->getLanguage() : '';
+        return $this->hasKey() ? $this->getTypePrefix() . $this->getFlexType() . '.' . $this->getKey() . '.' . $this->getLanguage() : '';
     }
 
     /**